% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade.r
\name{as_phyloxml}
\alias{as_phyloxml}
\alias{phyloxml}
\alias{as_phyloxml.phylo}
\title{Coerce into phyloXML}
\usage{
as_phyloxml(tree, ...)

\method{as_phyloxml}{phylo}(tree, name = "A phylogenetic tree",
  description = "Some description", xmlns = "http://www.phyloxml.org",
  digits = 20, ...)
}
\arguments{
\item{tree}{A phylogenetic tree.}

\item{...}{Further arguments to be passed to the method (see details).}

\item{name}{Character scalar. Name of the tree.}

\item{description}{Character scalar. Description of the tree.}

\item{xmlns}{Character scalar. Location of the default namespace
(see the \href{https://www.w3schools.com/xml/xml_namespaces.asp}{XML Namespaces}
section in w3schools).}

\item{digits}{Integer scalar. Precision to use when writing numbers.}
}
\value{
An XML document (see \link[xml2:read_xml]{xml2::read_xml})
}
\description{
Coerce phylogenetic trees to phyloXML documents, objects of class \code{xml_document}.
}
\details{
So far, this function has a method for \link[ape:as.phylo]{phylo} objects
from the \CRANpkg{ape} package. The following elements are supported:
\itemize{
\item Branch lengths as stored in \code{edge.length}
\item Labels, which are exported as \code{name}, as stored in \code{tip.label} and, if
available, \code{Node.label}.
\item Tree topology as stored in \code{edges}
}
}
\examples{

# Random tree from the ape package ------------------------------------------
set.seed(1)
x <- ape::rtree(5)
x

as_phyloxml(x)
}
\references{
Han M.V. and Zmasek C.M. (2009) "phyloXML: XML for evolutionary biology and
comparative genomics"  BMC Bioinformatics, 10:356 \doi{10.1186/1471-2105-10-356}
}
\seealso{
\link{validate_phyloxml}
}
