% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_to_phylo.R
\name{edgelist_to_phylo}
\alias{edgelist_to_phylo}
\title{Edgelist to phylo object}
\usage{
edgelist_to_phylo(x, edge_length = NULL, root_edge = NULL)
}
\arguments{
\item{x}{An object of class \link{matrix}.}

\item{edge_length}{A vector with branch lengths  (optional).}

\item{root_edge}{A numeric scalar with the length for the root node (optional).}
}
\description{
Edgelist to phylo object
}
