% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{data}
\name{phyloXML_ns}
\alias{phyloXML_ns}
\title{PhyloXML namespace
An object of class \code{xml_namespace}}
\format{
An object of class \code{xml_namespace} of length 2.
}
\usage{
phyloXML_ns
}
\description{
PhyloXML namespace
An object of class \code{xml_namespace}
}
\keyword{datasets}
