% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phyloxml.R
\name{read_phyloxml}
\alias{read_phyloxml}
\alias{phyloxml}
\alias{phyloxml-class}
\alias{read_phyloxml.character}
\alias{read_phyloxml.default}
\title{Read phyloXML files}
\usage{
read_phyloxml(x, ...)

\method{read_phyloxml}{character}(x, ...)

\method{read_phyloxml}{default}(x)
}
\arguments{
\item{x}{Either a character scalar (path to a phyloXML file) or an \code{xml} object.}

\item{...}{Further arguments passed to \link[xml2:read_xml]{xml2::read_xml}}
}
\value{
A list of class \code{phyloxml} in which each element represents a
tree:
A \code{data.frame} with the following elements:

\item{id}{Integer, the id of the node.}
\item{name}{Character, the name of the node (can be \code{NA}).}
\item{branch_length}{Numeric, the length of the branch (can be \code{NA}).}
\item{iselaf}{Logical, whether it is leaf (tip) or not.}
\item{parent}{Integer, the id of the parent node.}
}
\description{
Read phyloXML files
}
\details{
\code{phyloxml} objects can be coerced as \link[ape:multiphylo]{ape::multiphylo} using the
\link{phyloxml2phylo} function.
}
\examples{
library(ape)
set.seed(1)

x <- rtree(10)
x2 <- write_phyloxml(x)
read_phyloxml(x2)
}
