% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{validate_phyloxml}
\alias{validate_phyloxml}
\title{Validate a \code{phyloXML} object using the phyloxml.xsd schema.}
\usage{
validate_phyloxml(x, ...)
}
\arguments{
\item{x}{Either an object of class \code{xml_document} or a path to a file.}

\item{...}{Further options passed to \code{\link[xml2:read_xml]{xml2::read_xml()}}.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A wrapper of \link[xml2:xml_validate]{xml2::xml_validate}. Performs validation of phyloXML
objects using the phyloxml.xsd schema.
}
\details{
The schema can be found in \Sexpr{system.file("phyloxml/phyloxml.xsd", package="rphyloxml")}.
The latest version of the schema can be downloaded from \url{http://www.phyloxml.org/}
}
