% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_phyloxml}
\alias{write_phyloxml}
\alias{write_phyloxml.phylo}
\title{Save a phylo tree as phyloXML}
\usage{
write_phyloxml(
  tree,
  file = "",
  overwrite = FALSE,
  read_options = list(),
  write_options = list(),
  events = NULL,
  ...
)

\method{write_phyloxml}{phylo}(
  tree,
  file = "",
  overwrite = FALSE,
  read_options = list(),
  write_options = list(),
  events = NULL,
  name = "A phylogenetic tree",
  description = "Some description",
  xmlns = "http://www.phyloxml.org",
  digits = 20,
  ...
)
}
\arguments{
\item{tree}{A phylogenetic tree.}

\item{file}{Character scalar. When equal to \code{""}, the function only prints the xml
document and returns the object invisibly. Otherwise, it will try to write the xml
file to the specified path using \code{\link[xml2:write_xml]{xml2::write_xml()}}.}

\item{overwrite}{When \code{FALSE}, the function will check first whether \code{file} exists,
and if so, return with an error.}

\item{read_options, write_options}{List of arguments passed to \link[xml2:read_xml]{xml2::read_xml} and
\link[xml2:write_xml]{xml2::write_xml} respectively.}

\item{events}{A \link{data.frame} with events. It must be \code{nrow(events) = Nedge(tree)}
(see examples).}

\item{...}{Further arguments to be passed to the method (see details).}

\item{name}{Character scalar. Name of the tree.}

\item{description}{Character scalar. Description of the tree.}

\item{xmlns}{Character scalar. Location of the default namespace
(see the \href{https://www.w3schools.com/xml/xml_namespaces.asp}{XML Namespaces}
section in w3schools).}

\item{digits}{Integer scalar. Precision to use when writing numbers.}
}
\value{
An XML document (see \link[xml2:read_xml]{xml2::read_xml})
}
\description{
Coerce phylogenetic trees to phyloXML documents, objects of class \code{xml_document}.
}
\details{
So far, this function has a method for \link[ape:as.phylo]{phylo} objects
from the \code{ape} package. The following elements are supported:
\itemize{
\item Branch lengths as stored in \code{edge.length}
\item Labels, which are exported as \code{name}, as stored in \code{tip.label} and, if
available, \code{Node.label}.
\item Tree topology as stored in \code{edges}
}
}
\examples{

# Random tree from the ape package ------------------------------------------
set.seed(1)
x <- ape::rtree(5)
x

write_phyloxml(x)
}
\references{
Han M.V. and Zmasek C.M. (2009) "phyloXML: XML for evolutionary biology and
comparative genomics"  BMC Bioinformatics, 10:356 \url{doi:10.1186/1471-2105-10-356}
}
\seealso{
\link{validate_phyloxml}
}
